# Limit number of files in TestProjectGenerator.TestFunction_Speed to 5000