Logic for excluding duplicate files in `VSProjectGenerator` has O(n^2) complexity. This leads to apparent hanging of `TestFunction_Speed` if the user has large amount of extra files in the `Code` directory. In my case I had several copies of corpus for BFFFuzzer inside `Code` directory, about 90000 files in total.

To prevent confusion (it's not obvious what is the cause of hang) limit number of files passed to `VSProjectGenerator::AddFiles` to 5000, that many files can be handled in a reasonable time.