@ffulin 2018-06-25
I went ahead and fixed the performance regression (https://github.com/fastbuild/fastbuild/commit/03a7d87362dad5f69f2fffde8461faf7888fa28f). The performance is now almost the same as before in Release, and is significantly faster in Debug builds. I stressed it out with 90,000 files:

| | Old | Current | New |
| --- | --- | --- | --- |
| Debug | 66.6s | > 10 minutes* | 7.1s |
| Release | 3.75s | 448s | 4.0s |

\* I didn't bother waiting for it to finish :)

Thanks for letting me know about the regression. I'll leave this pull request open until the next version is released, so v0.96 will remain overall about the same as before, but with the de-duplication in-tact.