@ffulin 2018-07-06
A version containing this fix has now been released (v0.96).